feat: show notifications bell only for following user
