+++ src/components/NotificationBellBtn/hooks.js
@@ -1,6 +1,8 @@
 import { useMemo } from 'react'
 import gql from 'graphql-tag'
 import { useQuery, useMutation } from '@apollo/react-hooks'
+import { useUser } from '../../stores/user'
+import { isInFollowers } from '../../pages/profile/follow/FollowBtn'
 
 const NOTIFICATIONS_FOLLOWERS_QUERY = gql`
   {
@@ -115,4 +117,36 @@ export const useNotificationToggle = targetUserId => {
   const disabledBtn = loading || toggleRequestSending
 
   return { toggle, isNotificationDisabled, disabledBtn }
+}
+
+export const useIsInFollowers = targetUserId => {
+  const { user: currentUser } = useUser()
+  const { data: followData = {} } = useFollowers()
+
+  const usersList = useMemo(
+    () => {
+      const { following2 } = followData
+      if (!following2 || !following2.users) {
+        return []
+      } else {
+        return following2.users.map(({ userId }) => ({ id: userId }))
+      }
+    },
+    [followData]
+  )
+
+  console.log('usersList', usersList)
+
+  const isInFollowersList = useMemo(
+    () => {
+      if (!currentUser) {
+        return false
+      }
+
+      return isInFollowers(usersList, targetUserId, currentUser.id)
+    },
+    [usersList, targetUserId, currentUser]
+  )
+
+  return isInFollowersList
 }
